Update 2012/tromp
Format and typo fix README.md.

Un-rot13 the how13 and make it how.md. The reversal of rot13 was
discussed on GitHub at Landon's request.

One link was fixed.

Added other links. I was sorely tempted to make two links for garbage
collector, garbage == Windows and collector == the actual garbage
collector :-) but I decided that that's not my place so I didn't do it
as much as I wanted to (and I dearly wanted to!).

In some places there probably should be links again but since they were
linked to earlier in the document I'm going to pass on it for now.

I believe but am not certain that with this commit with the exception of
2012/README.md that 2012 is finished at least with format fixes in
README.md files (unless I come across something when trying out the
entries later on).