Several updates/alt code to 2013/birken
The alt version is based on the author's suggestion but like I've done
with the others (sometimes based on author's suggestions, sometimes not)
where usleep() is used I made it configurable at compile time.

Added demo.sh which cycles through all the text files provided, noting
prior to running the command what is about to be run, to show them all.
Note that if someone were to put in an invalid text file in the
directory there might be a problem. Script removes (in message) the
'.txt$' and capitalises the word. In some cases it changes the text to
be two words. Script passes shellcheck. This script allows one to use
the alt version as well; the try section of the README.md shows this
happening.

Format fix and added links (quite a few - I have very fond memories of
the games from the 1980s!) to the games referenced.

Updated .gitignore.