Add table generator to 1984/mullender from author
This included adding the 'a.out.h' as not all systems have it and also
I wanted it as close to the original as possible so I got a copy
(https://raw.githubusercontent.com/dspinellis/unix-history-repo/Research-Release/usr/include/a.out.h) from
the fabulous Unix History repo
(https://github.com/dspinellis/unix-history-repo) which I highly
recommend anyone interested in the evolution of Unix to check out!

... and yes: the repo is under the account of Diomidis Spinellis who is
a previous IOCCC winner!

To generate the tool run `make gentab'. Note that keeping the short ints
short generates a lot of negative numbers but this might or might not be
a problem: the original mullender.c has a negative number for instance.

Updated .gitignore and README.md files.