Add initial tmp/author-wins.csv
As Landon noted on GitHub:

    One of the key elements that we are missing from the SQL command set
    is a way to tie a winning author handle to the entry handles..

    We think what is needed is an CVS file, perhaps, tmp/author-wins.csv
    that connects the author handle with the winning entries that they
    won.

    The proposed tmp/author-wins.csv will be used to generate the list
    of winning entries that a given author has submitted. This will be
    used to generate the per-author JSON files.

As I just wanted to start it and I've not given any thought whatever at
how to generate it (except very briefly yesterday morning) I've only for
now included what Landon did. Once I've generated it all it'll be sorted
maybe by year or maybe by surname: that's to be determined later but
this is a nice start at least.