Fix 1993/Makefile to compile 1993/ant
When the idea of having the winning entry directory names match the name
of the authors was decided to be a bad idea after all I forgot to update
the 1993 Makefile with Anthony Howe's entry so that it tried to cd to
1993/howe and not 1993/ant. This caused a compilation error and
prevented 1993/ant from being compiled.

Very few entries do not compile now though that does not mean that all
that compile work - yet at least. Some that don't compile never will:

- 1988/spinellis will never compile with clang. I added alt code for
clang.

- 1993/lmfjyh will not compile except for gcc == (or < ?) 2.3.3 as it
relies on a bug in gcc that was fixed in version 2.3.3. This entry
unfortunately creates a bogus file called:
	'";main(){puts("Hello World!");}char*C=".c' though at least it's
	excluded in .gitignore ('"*'). The fact this file is generated
	is still an issue though and perhaps the entry should be made
	alt code and the original could be an #error if gcc >= 2.3.3?

- 1994/westley is not supposed to compile as depending on the
compiler flags it plays a game!

- 2001/anonymous/anonymous.ten.c has to be compiled with -m32 and not
all platforms are able to do this (glaring at macOS here).

- Similarly to 2001/anonymous/anonymous.ten.c this too has to be
compiled with -m32 and not all platforms permit code to be compiled as
32-bit.

- 2004/gavin cannot be linked in some architectures / platforms like
macOS. I actually have a fix I came up with the other day that lets it
compile with macOS but it doesn't do anything either (I have one theory
why that is). I might add this to the repo just so it can at least
compile (via make alt). Even so this entry has some problems for modern
systems including architecture.

- One file in 2015/yang does not compile but I believe this to be the
Makefile. I'll fix this in a near future commit if that is indeed the
case.

All other entries compile (many did not but I fixed many - including
removing the need for -traditional-cpp; Landon also fixed a few and so
did Yusuke Endoh)! Not all of those work however (in all platforms or at
all). I have not tested all either. I am still in the process of this
but these are of lower priority for the moment (though I might address
some before the higher priority tasks if I don't feel up to said tasks
but do feel up to these). This means we've made tremendous progress in
issue #1!