Clean up 1993/lmfjyh on compilation error
Now if the compilation fails (it relies on a bug in gcc version < 2.3.3)
the bogus file is removed. It is not possible to check the gcc version
in the source code because the file is copied from lmfjyh.c in the
Makefile prior to the compilation. Well it's possible to check the gcc
version but it's too late at that point - the file would have already
been made :-)

This prevents the user form having to do quoting of the file (risking
making a mistake) or using -find . -inum ... -delete after obtaining the
inode (ls -li).