Add SQL files for json files
The two files - prizes.sql and winners.sql - are to be used for
verifying the CSV file for eventual json files. I do have the other SQL
files as well but I don't think they should be needed and some are out
of date (think URLs for instance). Having the files in the repo will
make it much easier for me when I look at them as they were elsewhere on
disk.

I put the files in tmp/sql as that felt the most appropriate place and
way to identify them.