Fix 2004/gavin/{kernel,fs.tar} commit
Landon pointed out some issues that I missed (thank you Landon .. I was
very tired and I hope that I got everything this time) in commit
c75806fb3360f02cde73ed830700cc38fb5a8ac0:

0.  The make clobber rule removes the git files 2004/gavin/kernel and
    2004/gavin/fs.tar.

1.  The 2004/gavin/fs.tar file contains the file "README.md" which, when
    untarred, will overwrite the git file 2004/gavin/README.md.

2.  The 2004/gavin/fs.tar file contains the file "gavin.c" which, when
    untarred, will overwrite the git file 2004/gavin/gavin.c.

3.  The 2004/gavin/fs.tar file is a tar file that could be related with
    just checking in the 3 files "prim", "sh" and "vi".

and I discovered another issue too namely that the tarball had the wrong
version of gavin.c in it as well (prior to the fix for clang).

Now the tarball is in the img/ subdirectory and the tarball contents (is
that tar ? :-) ) looks like:

    $ tar ft img/fs.tar
    fs/
    fs/gavin.c
    fs/vi
    fs/README.md
    fs/sh
    fs/prim

I added to the todo.md file that this file should be verified to match
the contents of 2004/gavin and I gave a command to do to test this - at
the end of working on this entry or maybe better yet right before this
repo is to be folded back into the winner repo.

Updated .gitignore so that the files that are in img/fs.tar are not
added to the repo. The tarball is not meant to be extracted anyway but
rather copied to a disk for usage of the entry. It will be extracted to
verify that the contents match that of the entry itself (2004/gavin
directory).