Add alt code for 1993/lmfjyh
Why? Because! :-)

Updated .gitignore.

Added the entry to bugs.md.

Slightly modified the note in the Makefile so that the make all rule
refers to make alt.

The README.md file was updated including succinctly giving the gist
(that might or might not be redundant :-) ) of the entry i.e. how it
worked (though the author also stated this).