Make progress on 2005/giljade with clang
I've actually figured out the problem and now the errors (on running the
program on its output) is down to 3 errors instead of I think it was 9.
I have not updated bugs.md but I will in time if I don't end up fixing
this (in the end I will most likely fix it or at least attempt to). A
look at the diff here might give a clue however. Still I do intend to
address this further to get it to work properly.