More twitter deprecation - 2019/mills
In the README.md in the judges' remarks there was a link not to
twitter.com but an archive of a certain person's tweet. However this
link is dead and I don't think it is worth using the Internet Wayback
Machine in this case due to if nothing else major divisiveness for the
US and the world full stop. The name is still there but no link. I'm not
even sure about that. Perhaps since twitter is entirely deprecated it
should be removed entirely but I don't know.