Add note about sockets lying about in 2019/duble
The entry appears to leave sockets lying about in the current working
directory. Although undesirable this is not a bug it's a feature.
Updated bugs.md and the README.md file as well as the Makefile.

A useful thing I thought of that might be good to look at for other
entries that show notes in the Makefile is that it might be good (and in
this entry I did do it) to have the note _after_ compilation as it will
be more likely seen. In the case of entries that might not compile with
clang (due to some defects in it) it will have to be done with either ||
or else printed before the compilation (for instance in some of these we
have to say that one should use make alt for clang).

This will be added to the todo.md in a future commit.