Fix details about 2019/duble sockets
Of course the globs do not have a = at the end of the line; that's from
an aliased ls that shows different characters at EOL depending on the
file type.

I also explained how to delete these files assuming that glob pattern.

In the README.md this was done in the new INABIAF section which is where
it belonged in the first place.