Undo 6b6e057e96d7f574904ffece8328916b522c3693
I meant to do a commit --amend as I realised I had another change to
make but then I forgot and went ahead and did another commit. This means
I have to do this change in two commits. This will undo the commit noted
and the next will be the proper change though I now do not know exactly
what else had to change: it might have been done in the commit that
followed (I think it was to do with sockets but maybe not). Nevertheless
the log was incomplete so I want to fix that.