Various updates in todo.md
What prompted this update is a 'problem' I noticed in 2019/duble: it
leaves sockets lying round in the current working directory but it might
not be safe for make clobber to delete them: involves '.' files and we
don't know if the glob will be different in different systems. In macOS
it appears to be '^\.[A-Z].*' (which in commit
6b6e057e96d7f574904ffece8328916b522c3693 I made a mistake saying it
ended with '+*' which is obviously very wrong).

The README.md and bugs.md does show how to delete that pattern but it is
TBD if make clobber should do this.