Fix command to find sockets in 2019/duble
The grep expression in particular was incorrect. This commit fixes the
problem. Note that it doesn't look for lower case letters as there are
legitimate dot files in the repo with lower case letters. Hopefully no
system will leave lower case file sockets lying about (though for all I
know it's the entry deciding it - didn't look into it).