Remove extra thank yous + links in thanks file
As there already are thanks at the top there seems to be no reason to
have the extra thanks for each entry. This was due to yank paste from
the README.md files which naturally made the change much more efficient.

I also removed the links to both Yusuke and myself in the entries
themselves as they're at the top of the file.

I also (because I am very proud) added some other things I did at the
top of the honour list.