Fixes and improvement in 1984/anonymous display
The tattoo image now has the code at the top of it (separate but in the
same image) so one can see both next to each other.

The README.md had some slight improvements in wording and so did the
Makefile. Also with the Makefile I tried to make it clearer that the alt
entry might not work on some systems. This same style (or something like
it) could be done for other entries as well.