Fixes in 1989/westley
Try and improve the note during compilation that some versions will not
work with some compilers.

Make the other versions alt code. Running make alt will stop on the
first error but I made it alt targets so that make clobber will remove
the compiled files. In order to make it actual targets the Makefile
would have to do more to generate the files and this might or might not
be better.

Format fixes and a typo fix in README.md.