Improve how 1989/westley alt versions are built
The Makefile now has the other versions as alt targets. Running make
ver0 will generate ver0.c and then try compiling it. Running make ver1
will do the same for ver1.c and running make alt will do this on ver0,
ver1, ver2 and ver3. Error in compilation in these rules are not fatal
so that make alt can continue. This is important for clang.

I have some other enhancements on the presentation of this entry but
they have to be reworked and decided upon (it might be that it's not
necessary with the alt rule but then the README.md file really ought to
be updated too which it will be in another commit).

It might be that making these alt targets is not even the right approach
but for now it'll do as at least it's easier to generate and compile the
individual files even if one does not use make alt. This is all TBD
later.