Fix 2004/burley
There were two things that had to change with my previous fix (getting
it to compile), one in code and one in the Makefile.

Instead of doing ',1' after the calls to longjmp() do ',0'. The way it
originally was is in binary expressions but since the function returns
void I made the result be 0 rather than 1 and along with the other
change it works (though maybe it would work with 1 anyway but it seems
not to be necessary and I don't feel like testing it now).

The Makefile change is that it won't work right with the optimiser
enabled so that was disabled.

I removed this entry from bugs.md (which was listed twice by accident -
and some other things were too) and updated it in the
thanks-for-fixes.md file as well.