Update faq.md
I added an entry on how to get sound entries to work under macOS (well
some already will work but some need SoX and I explained two ways to
install it - MacPorts and Homebrew).

I added an entry explaining where to find a list of entries with known
issues or things that appear to be issues but are rather features.

I also elaborated on some entries and I moved the terminal sanity (or
insanity :-) ) entry closer to the top as it seems like it is better to
help restore terminal sanity closer to the top as a messed up terminal
is very annoying.