Fix 2004/jdalbec for gcc
GCC objected to the '##' and some other cpp directives that were being
used.

This unfortunately did mess up the formatting of the some of the code
but I tried to preserve the art and the periodic table of elements
formatting (this was preserved but whether I got all the art or not
preserved I do not know) as these parts seemed too important to not
preserve.