Fix 2015/endoh3 for linux
There was a problem with duplicate symbols of main but using the option
-fcommon in the compiler flags solves this problem. It was not an issue
with macOS as the default behaviour (I guess default anyway).

With this commit I believe that all entries (except for two that are not
supposed to compile) compile in linux. They haven't all been tested but
all but two should compile in linux if one has the appropriate libraries
for those that use other libraries.