Fix funny problem in alt target in 1989/westley
In linux running make ver1 and make ver2 and make ver3 (not sure about
ver0 now) there was a strange error when running the rules that meant
the code was not compiled. Thinking on it just now it very likely was
the exit code of the program causing the problem but the file was
generated so I added a '-' to the line so that the rule will proceed to
try and compile the code.