Change bug status of 1996/august
On another reading of one of the files it occurred to me that he hanging
I observed with clang (after fixing it for gcc) was actually documented
by the judges in the first place. They noted that some compilers compile
the code into an infinite loop and that is indeed what clang does. Thus
because it's documented it's a feature not a bug.