Format and typo fixes to 1984 README.md files
I paid a lot of attention to 1984/mullender partly because it is one of
Landon's all time favourite entries to this day and partly because it
felt like it was a good idea especially as there is a lot of
instructional value in the author's remarks that I located a while back.
However this did take quite some time so I probably won't do that for
many more entries if any more as it will slow things down tremendously.
There was an unfortunate typo fix in this README.md too: a typo in the
file, possibly a typo from the copy paste (think I copied from an image
but it might have been a pdf too in which case the original person
adding the remarks in the book cited made a typo / copy-pasted the
information from the author), which could cause confusion when reading
the remarks.