Add missing files to 2015/endoh3
After emailing Yusuke he provided me with the missing files which I have
added to the repo. I removed this part from the thanks (I had noted that
there were some files missing) and removed the entry from the bugs.md
file.

Yusuke offered a hint and through this I noticed the main[] in the
README.md (I haven't yet looked at year 2015 in detail). It is what
generates the fun.tar.gz file (which is now in the .gitignore file as
the files within - except for the README.markdown there - are now in the
repo). I put this code in fun.c and added a rule to the Makefile to
build it (and it is built by default with make all). I also added to the
Makefile the rule to build (also by default) bf2pdp11.c. The README.md
notes how to use fun.c but it does not show how to use bf2pdp11 as I'm
not sure of the main use yet (I guess brainfuck to pdp11 which might be
useful with the brainfuck files added but that's all I can guess at this
time).