Move src/sgit.README.md -> src/sgit/README.md
Since it is now in its own directory it seems better to have a regular
README.md file. It might be better to just refer to the GitHub repo so
it's not necessary to have to remember to update it whenever I update
the master copy but since I have to do the same with sgit itself maybe
it doesn't matter. For now this will do.