Sync sgit.sh to be sgit from sgit repo
The repo is at https://github.com/xexyl/sgit.

Changed header comments in Makefile, noting that Landon made the
Makefile (here) but that the most up to date copy of the tool can be
found at my repo noted above. This is to make it easier in case there
comes a time where I am unable to update it here right away and then
forget to do it later after updating the repo (though it seems unlikely
that this will ever need to be done I'd like this in as a safety net).