Add alt code to 1986/marshall
Add the code before the fixes to the entry that prevented it from
working right depending on the compiler and optimiser (where having the
optimiser with one compiler would let it work with one and not work with
the other but having it disabled would reverse this in the same way) and
documented the ways the compilers conflicted for instructional value and
entertainment.

Updated .gitignore.