added initial draft of manifest.
The `manifest.numbers` contains information about files, that
must already exist.  See `README.md` for details.

The `manifest.csv` is generated from `manifest.numbers`.
See `README.md` for details on how it is generated.

The `missing-manifest.numbers` contains information about files,
that may not yet exist because a tool needs to create it.

The `missing-manifest.csv` is generated from `manifest.numbers`.
See `README.md` for details on how it is generated.

Updated `example.dot_winner.json` as per fields in
the `manifest.numbers` spreadsheet.

The `file_list.txt` contains paths of all files that should
already exist, or will be created by a tool.

The `winner_dir.txt` contains a list of paths to
winning directories.

Updated `README.md`.

NOTE: The `manifest.numbers` and `missing-manifest.numbers` are
initial draft files and will be worked in and modified.
In particular, empty cells represent values that need to be
added, or replaced with null if not such field file is needed.

NOTE: Lines from `missing-manifest.numbers` will be moved
into `manifest.numbers` as they are created.  Eventually
the `missing-manifest.numbers` spreadsheet will be come empty
and, along with `missing-manifest.csv` will be removed.

NOTE: Some files, such as `.id` files are temporary,
are not listed in any contents of the above files,
and will be removed.