Minor fixes in 1990/tbr
Remove note in Makefile that this entry does not compile with modern
compilers. I fixed that quite a while ago but neglected to remove the
message.

The README.md I changed 'i/o' to 'I/O' and I removed the reference to
the original version with gets() from the archive as it's now in the
tbr.orig.c file.