Add entries to FAQ + minor fixes to thanks file
The following entries were added:

0. Makefile rules.
1. 32-bit entries not working in 64-bit systems.
2. Some entries not working in macOS either uncompilable or
non-functional.
3. Entries that cannot be compiled with clang.

For (3) this especially is problematic for macOS as even /usr/bin/gcc is
clang (as has been noted many times before this but not in the FAQ).

At least one typo fix was also made and at least one entry that already
existed was updated a bit.

The fixes to the thanks file are related to the FAQ entries added.