Improve my original fix to 1987/wall
This improvement redefines gets() to use fgets() with the correct args
so that the code can refer to gets() to look more like the original
code.

Note that the prototype for gets() that is in the original code cannot
be there because of conflicting types of fgets() and a couple other
errors that arise from redefining gets(). This fix is part of issue #936
which with this commit is more than halfway complete.